\name{wsrpart-package}
\alias{wsrpart-package}
\docType{package}
\title{

  Weighted sub-space rpart forests

}
\description{

  Build a forest of rpart decision trees, randomly sampling both the
  observations and the variables. Variables are randomly sampled weight
  by a probability calcaulted from the variable's correlation to the
  output.
  
}
\details{
\tabular{ll}{
Package: \tab wsrpart\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-22\cr
License: \tab GPL v3\cr
}

}
\author{
  Zhalama and Graham Williams(SIAT)

  Maintainer: Graham Williams <Graham.Williams@togaware.com>

}
